Add files via upload
The server can serve any number of sessions at the same time. Every session deals with two 
players. The client can be conveyed to run as a Java applet. To run a client as a Java applet from a Web program, the server must run from a Web server. 
The TicTacToeRules interface characterizes the constants shared by all the classes in the 
project. Each class that utilizes the constants needs to actualize the interface. Typically, characterizing constants in an interface is a typical practice in Java.